chore: Move transformer to its own file
